// f/DataIngestion/src/app/workspace-connect/core/workspace-connect.types.ts
export interface SharepointEnvironment {
  production: boolean;
  apiBaseUrl: string;
  apiVersion: string;
  ownerKey: string;
  defaultLibraryName: string;
  tenantName: string;
  clientId: string;
  authority: string;
  hostName: string;
  siteName: string;
  libraryName: string;
  userGraphScopes: string[];
}

export type SharepointUserConfig = Pick<
  SharepointEnvironment,
  'tenantName' | 'clientId' | 'authority' | 'hostName' | 'siteName' | 'libraryName' | 'userGraphScopes'
>;

export const SHAREPOINT_ENV = new InjectionToken<SharepointEnvironment>('SHAREPOINT_ENV');